Wrote an inliner module because inlining code was getting too sizeable (had to make some compiler.rs data structures public)
