+++ som-interpreter-bc/src/lib.rs
@@ -30,6 +30,9 @@ pub mod universe;
 /// Facilities for manipulating values.
 pub mod value;
 
+/// Inlining some calls to a select few builtin functions for sizeable perf gains.
+pub mod inliner;
+
 /// A strong and owning reference to an object.
 pub type SOMRef<T> = Rc<RefCell<T>>;
 /// A weak reference to an object.
